% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hist.R
\name{plot_hist}
\alias{plot_hist}
\title{Plot histograms for all numerical columns in the given dataframe}
\usage{
plot_hist(input_df, density = FALSE, exclude = c(""), title = "Histogram")
}
\arguments{
\item{input_df}{The input dataframe}

\item{density}{The boolean indicating whether or not to plot the density.}

\item{exclude}{The columns not to include in the output.}

\item{title}{The title of the histograms}
}
\value{
ggplot2 object
}
\description{
plot histograms for all numerical columns in the given data frame.
}
\examples{
plot_hist(mtcars, exclude = c(""))
}
