% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scatter.R
\name{plot_scatter}
\alias{plot_scatter}
\title{Scatter Plot Function}
\usage{
plot_scatter(data, xcol, ycol, color, title = NULL, size = 20)
}
\arguments{
\item{data}{A data frame or a tibble.}

\item{xcol}{A numerical column .}

\item{ycol}{A numerical column.}

\item{color}{A categorical column.}

\item{title}{An optional character variable spefified by users.}

\item{size}{An optional integer variable spefified by users for the text size of all labels.}
}
\value{
A scatter plot.
}
\description{
Takes a dataframe and returns a scatterplot of chosen numeric features in the dataset
}
\examples{
plot_scatter(iris, xcol = Sepal.Length, ycol = Sepal.Width,
color = Species, title = "Sepal.Length vs Sepal.Width", size = 20)
}
