% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scatter.R
\name{scatter_plot}
\alias{scatter_plot}
\title{Scatter Plot Function}
\usage{
scatter_plot(
  data,
  x,
  y,
  color = "green",
  title = "Scatter plot",
  plot_width = 300,
  plot_height = 300
)
}
\arguments{
\item{data}{A data frame or a tibble.}

\item{x}{A character variable.}

\item{y}{A character variable.}

\item{color}{A color specified by users.}

\item{title}{An optional character variable spefified by users.}

\item{plot_width}{An optional integer variable spefified by users.}

\item{plot_height}{An optional integer variable spefified by users.}
}
\value{
A scatter plot.
}
\description{
Takes a dataframe and returns a scatterplot of chosen numeric features in the dataset
}
